+++ CocoJumper.Base/EventModels/SearchResultEvent.cs
@@ -1,16 +1,16 @@
-﻿using System.Collections.Generic;
-using System.Linq;
+﻿using CocoJumper.Base.Events;
+using System.Collections.Generic;
 
-namespace CocoJumper.Base.Events
+namespace CocoJumper.Base.EventModels
 {
     public class SearchResultEvent
     {
-        public List<SearchEvent> SearchEvents;
-        public bool IsHighlightDisabled;
-
         public SearchResultEvent()
         {
             SearchEvents = new List<SearchEvent>();
         }
+
+        public bool IsHighlightDisabled { get; set; }
+        public List<SearchEvent> SearchEvents { get; set; }
     }
 }
