+++ CocoJumper/Commands/CocoJumperOptions.cs
@@ -7,55 +7,50 @@ namespace CocoJumper.Commands
     public class CocoJumperOptions : DialogPage
     {
         private const string GeneralCategory = "General";
-        private int _timerInterval = 250;
-        private int _limitResults = 50;
-        private int _automaticallyExitInterval = 5000;
-        private bool _jumpAfterChoosedElement = false;
-        private bool _disableHighlightForSingleSearch = false;
-        private bool _disableHighlightForMultiSearch = false;
-        private bool _disableHighlightForSingleHighlight = false;
 
-        [Category(GeneralCategory)]
-        [DisplayName("Limit results")]
-        [Description("Limts results that are rendered on one page.")]
-        [DefaultValue(50)]
-        public int LimitResults { get => _limitResults; set => _limitResults = value; }
-
-        [Category(GeneralCategory)]
-        [DisplayName("Timer interval(ms)")]
-        [Description("Determines how much ms must pass before rendering components, counting from last key press.")]
-        [DefaultValue(250)]
-        public int TimerInterval { get => _timerInterval; set => _timerInterval = value; }
+        public event EventHandler Saved;
 
         [Category(GeneralCategory)]
         [DisplayName("Automatically exit after(ms)")]
         [Description("Determines how much ms must pass before logic will automaticly exit.")]
         [DefaultValue(5000)]
-        public int AutomaticallyExitInterval { get => _automaticallyExitInterval; set => _automaticallyExitInterval = value; }
-
-        [Category(GeneralCategory)]
-        [DisplayName("Jump after choosed element")]
-        [Description("If set to True, logic will move caret to the end of choosed element.")]
-        [DefaultValue(false)]
-        public bool JumpAfterChoosedElement { get => _jumpAfterChoosedElement; set => _jumpAfterChoosedElement = value; }
+        public int AutomaticallyExitInterval { get; set; } = 5000;
 
         [Category(GeneralCategory)]
         [DisplayName("Disable highlight for multisearch")]
         [Description("If set to True, logic will not render any highlighting components with may slowdown VisualStudio.")]
         [DefaultValue(false)]
-        public bool DisableHighlightForMultiSearch { get => _disableHighlightForMultiSearch; set => _disableHighlightForMultiSearch = value; }
+        public bool DisableHighlightForMultiSearch { get; set; } = false;
 
         [Category(GeneralCategory)]
         [DisplayName("Disable highlight for single search with select")]
         [Description("If set to True, logic will not render any highlighting components with may slowdown VisualStudio.")]
         [DefaultValue(false)]
-        public bool DisableHighlightForSingleHighlight { get => _disableHighlightForSingleHighlight; set => _disableHighlightForSingleHighlight = value; }
+        public bool DisableHighlightForSingleHighlight { get; set; } = false;
 
         [Category(GeneralCategory)]
         [DisplayName("Disable highlight for single search")]
         [Description("If set to True, logic will not render any highlighting components with may slowdown VisualStudio.")]
         [DefaultValue(false)]
-        public bool DisableHighlightForSingleSearch { get => _disableHighlightForSingleSearch; set => _disableHighlightForSingleSearch = value; }
+        public bool DisableHighlightForSingleSearch { get; set; } = false;
+
+        [Category(GeneralCategory)]
+        [DisplayName("Jump after choosed element")]
+        [Description("If set to True, logic will move caret to the end of choosed element.")]
+        [DefaultValue(false)]
+        public bool JumpAfterChoosedElement { get; set; } = false;
+
+        [Category(GeneralCategory)]
+        [DisplayName("Limit results")]
+        [Description("Limts results that are rendered on one page.")]
+        [DefaultValue(50)]
+        public int LimitResults { get; set; } = 50;
+
+        [Category(GeneralCategory)]
+        [DisplayName("Timer interval(ms)")]
+        [Description("Determines how much ms must pass before rendering components, counting from last key press.")]
+        [DefaultValue(250)]
+        public int TimerInterval { get; set; } = 250;
 
         public override void SaveSettingsToStorage()
         {
@@ -69,7 +64,5 @@ namespace CocoJumper.Commands
 
             Saved?.Invoke(this, EventArgs.Empty);
         }
-
-        public event EventHandler Saved;
     }
 }
